\name{dfHclust}
\alias{dfHclust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
demonstrate shiny utilities for hierarchical clustering
}
\description{
demonstrate shiny utilities for hierarchical clustering
}
\usage{
dfHclust(df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{
data.frame instance
}
}
\details{
shiny and cluster package are used to generate components of an
interactive hierarchical clustering interface
}
\value{
runs a shiny app at the browser
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
VJ Carey <stvjc@channing.harvard.edu>
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
(dfHclust(mtcars))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }% use one of  RShowDoc("KEYWORDS")
