\name{ph525x-package}
\alias{ph525x-package}
\alias{ph525x}
\docType{package}
\title{
\packageTitle{ph525x}
}
\description{
\packageDescription{ph525x}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ph525x}
\packageIndices{ph525x}

Software that can be used with the ph525x course that is not
yet in a CRAN or Bioconductor package.

}
\author{
\packageAuthor{ph525x}

Maintainer: \packageMaintainer{ph525x}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
ph525x() # verify installation
}
